+++ packages/iceworks-client/src/stores/index.js
@@ -1,11 +1,11 @@
 import Icestore from 'icestore';
 import projects from './projects';
-import currentProject from './currentProject';
+import project from './project';
 import materials from './materials';
 
 const icestore = new Icestore();
 icestore.registerStore('materials', materials);
-icestore.registerStore('project', currentProject);
+icestore.registerStore('project', project);
 icestore.registerStore('projects', projects);
 
 export default icestore;
